+++ src/main/java/roomescape/dto/PaymentRequest.java
@@ -1,8 +1,10 @@
 package roomescape.dto;
 
+import java.math.BigDecimal;
+
 public record PaymentRequest(String paymentKey,
                              String orderId,
-                             int amount) {
+                             BigDecimal amount) {
     public static PaymentRequest from(ReservationWithPaymentRequest reservationWithPaymentRequest) {
         return new PaymentRequest(reservationWithPaymentRequest.paymentKey(),
                 reservationWithPaymentRequest.orderId(),
